requirements.txt
aerich==0.7.2
asyncpg==0.28.0
email-validator==2.0.0.post2
fastapi==0.100.0
passlib==1.7.4
pydantic<2.0.0
python-jose==3.3.0
tortoise-orm==0.19.3
uvicorn==0.23.1
firebase-admin==6.2.0
google-api-python-client==2.95.0
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
python-multipart
